chore(metadata): Adds endpoint for data shapes
This is the first iteration on the data shapes endpoint that returns
JSON Schema of input and output shapes. As it stands it does not take
into the account the action itself, which currently stands to reason if
the action properties that denote what data is requested (name of the
stored procedure or Salesforce object type) is all that is required to
determine the shapes of data.
For Salesforce this is only partially true, as action will be needed
to distinguish from actions that operate on single object and actions
that operate on multiple objects (such as querying data).